Move blockchain to primitives
