+++ src/concurrency/ConsensusController.h
@@ -8,7 +8,7 @@
 
 #include "types/DevvMessage.h"
 #include "consensus/UnrecordedTransactionPool.h"
-#include "consensus/blockchain.h"
+#include "primitives/blockchain.h"
 #include "consensus/chainstate.h"
 
 namespace Devv {
